pyproject.toml
[tool.poetry]
name = "tomatex"
version = "0.1.0"
description = "Pomodoro timer backend"
authors = ["Gilson Filho <me@gilsondev.in>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pytest-django = "^4.1.0"
ipython = "^7.19.0"
ipdb = "^0.13.4"
freezefrog = "^0.4.1"
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
